Split master/node code from subnets.go into separate files
